+++ Cargo.toml
@@ -5,3 +5,6 @@ authors = ["bumblepie <luke@bumblepie.space>"]
 
 [dependencies]
 serial = "0.4.0"
+grbl = { path = "grbl" }
+
+[workspace]
